+++ internal/cmd/wiki/wiki.go
@@ -24,6 +24,7 @@ func NewWikiCmd() *cobra.Command {
 	cmd.PersistentFlags().StringVar(&wiki, "wiki", "", "URL of wikis api.php")
 	cmd.PersistentFlags().StringVar(&wikiUser, "user", "", "A user to interact using")
 	cmd.PersistentFlags().StringVar(&wikiPassword, "password", "", "Password of the user to interact with")
+	cmd.MarkFlagRequired("wiki")
 
 	return cmd
 }